move copy_assets, generate_tsconfig, create_app etc into new sync module
