+++ packages/kit/src/core/sync/create_manifest_data/test/static/foo.txt
