+++ packages/kit/src/core/utils.js
@@ -1,11 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 import colors from 'kleur';
-import { copy } from '../utils/filesystem.js';
 import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
 
 export const get_runtime_path = process.env.BUNDLED
 	? /** @param {import('types').ValidatedConfig} config */ (config) =>
@@ -16,25 +12,6 @@ export const get_runtime_path = process.env.BUNDLED
 function posixify_path(str) {
 	const parsed = path.parse(str);
 	return `/${parsed.dir.slice(parsed.root.length).split(path.sep).join('/')}/${parsed.base}`;
-}
-
-/** @param {string} dest */
-export function copy_assets(dest) {
-	if (process.env.BUNDLED) {
-		let prefix = '..';
-		do {
-			// we jump through these hoops so that this function
-			// works whether or not it's been bundled
-			const resolved = path.resolve(__dirname, `${prefix}/assets`);
-
-			if (fs.existsSync(resolved)) {
-				copy(resolved, dest);
-				return;
-			}
-
-			prefix = `../${prefix}`;
-		} while (true); // eslint-disable-line
-	}
 }
 
 function noop() {}
@@ -86,7 +63,7 @@ export function resolve_entry(entry) {
 	return null;
 }
 
-/** @param {import('./create_app/index.js').ManifestData} manifest_data */
+/** @param {import('./sync/sync.js').ManifestData} manifest_data */
 export function get_mime_lookup(manifest_data) {
 	/** @type {Record<string, string>} */
 	const mime = {};
